Adaptation to zephyr, more compile options
Added:

- WITH_RPUTILS_STATIC for linking statically with librp-utils
- WITH_BINDINGS_V3 for including V3 binding compatibility
- WITH_WSJ1 for handling WSJ1 protocol
- WITH_WSAPI for handling WSAPI protocol
- WITHOUT_JSON_C for removing use of libjson-c

+++ src/libafb/sys/x-uio.c
@@ -49,7 +49,7 @@ ssize_t readv(int fd, const struct iovec *iov, int iovcnt)
 			} while (rc < 0 && errno == EINTR);
 			if (rc <= 0)
 				return result ? result : rc;
-			result += (size_t)rc;
+			result += rc;
 			size -= (size_t)rc;
 			base = ((char*)base) + (size_t)rc;
 		}
@@ -75,7 +75,7 @@ ssize_t writev(int fd, const struct iovec *iov, int iovcnt)
 			} while (rc < 0 && errno == EINTR);
 			if (rc <= 0)
 				return result ? result : rc;
-			result += (size_t)rc;
+			result += rc;
 			size -= (size_t)rc;
 			base = ((char*)base) + (size_t)rc;
 		}
